updated bootstrap of config, and updated shebang
Shebang has been updated to acoount for different locations of lua.
This fixes #13.

Updated bootstrapping of config struct. This is now all handled in
configload module in lib.

Updated how libraries are found. This means that one has to set the
LUA_PATH=<path>/<of>/<gpm> before running the first initialize.

A fix will be made to write this load into the gpm module file,
which can then be used afterwards (such that it is only the very first
time you run gpm, that you have to manually set the LUA_PATH).
(maybe I will make a script for this at some point...).